+++ assets/js/util/tracking/constants.js
@@ -1,0 +1,31 @@
+/**
+ * Constants related to even tracking.
+ *
+ * Site Kit by Google, Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * Data layer global used for internal/private Site Kit data.
+ *
+ * @private
+ */
+export const DATA_LAYER = '_googlesitekitDataLayer';
+
+/**
+ * gtag script identifier.
+ *
+ * @private
+ */
+export const SCRIPT_IDENTIFIER = 'data-googlesitekit-gtag';
