+++ assets/js/util/tracking/createDataLayerPush.js
@@ -2,7 +2,7 @@
 /**
  * Internal dependencies
  */
-import { DATA_LAYER } from './createEnableTracking';
+import { DATA_LAYER } from './constants';
 
 /**
  * Returns a function which, when invoked will initialize the dataLayer and push data onto it.
