+++ assets/js/util/tracking/index.test.js
@@ -3,7 +3,7 @@
  * Internal dependencies
  */
 import createTracking from './createTracking';
-import { DATA_LAYER } from './createEnableTracking';
+import { DATA_LAYER } from './constants';
 import {
 	disableTracking,
 	enableTracking,
